src/requirements.txt
fastapi==0.61.1
ldap3==2.8.1
uvicorn==0.11.8
pydantic==1.6.1
requests==2.24.0
pycryptodome==3.9.9
python-jenkins==1.7.0
sqlalchemy==1.3.22
PyMySQL==0.9.3
Jinja2==2.11.2
aiofiles==0.6.0
python-multipart==0.0.5
passlib==1.7.2
PyJWT==1.7.1
bcrypt==3.2.0
suds-py3==1.4.4.1
paramiko==2.7.2
qrcode==6.1
pyotp==2.6.0